# api/openapi/http-api.yaml
openapi: 3.0.2
servers:
  - url: /v1
info:
  version: 1.0.0
  title: OpenEverest schema
  description: |
    # Authentication
    All requests to Everest API require `Authorization: Bearer <token>` header with a valid token in plain-text.
    
    The token can be obtained by using `everestctl token reset` which resets the token and prints it to the screen.
...
security:
  - BearerAuth: []
paths:
...
  '/clusters/{cluster}/instance-presets':
    x-everest-resource-name: instance-presets
    get:
      tags:
        - Preset
      summary: List instance presets
      description: |
        This API lists all instance presets in the specified cluster.
        Optionally filter by provider.
      operationId: listInstancePresets
      parameters:
        - name: cluster
          in: path
          required: true
          description: The name of the cluster
          schema:
            type: string
        - name: provider
          in: query
          required: false
          description: Filter presets by provider name
          schema:
            type: string
      responses:
        '200':
          description: List of instance presets
          content:
            application/json:
              schema:
                $ref: 'crds.gen.yaml#/components/schemas/InstancePresetList'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - Preset
      summary: Create instance preset
      description: |
        This API creates a new instance preset in the specified cluster.
      operationId: createInstancePreset
      parameters:
        - name: cluster
          in: path
          required: true
          description: The name of the cluster
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: 'crds.gen.yaml#/components/schemas/InstancePreset'
      responses:
        '201':
          description: Instance preset created
          content:
            application/json:
              schema:
                $ref: 'crds.gen.yaml#/components/schemas/InstancePreset'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  '/clusters/{cluster}/instance-presets/{name}':
    x-everest-resource-name: instance-presets
    get:
      tags:
        - Preset
      summary: Get instance preset
      description: |
        This API gets the instance preset specified by the `name` in the specified `cluster`.
      operationId: getInstancePreset
      parameters:
        - name: cluster
          in: path
          required: true
          description: The name of the cluster
          schema:
            type: string
        - name: name
          in: path
          required: true
          description: The name of the instance preset
          schema:
            type: string
      responses:
        '200':
          description: Instance preset details
          content:
            application/json:
              schema:
                $ref: 'crds.gen.yaml#/components/schemas/InstancePreset'
        '404':
          description: Instance preset not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - Preset
      summary: Update instance preset
      description: |
        This API updates the instance preset specified by the `name` in the specified `cluster`.
      operationId: updateInstancePreset
      parameters:
        - name: cluster
          in: path
          required: true
          description: The name of the cluster
          schema:
            type: string
        - name: name
          in: path
          required: true
          description: The name of the instance preset
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: 'crds.gen.yaml#/components/schemas/InstancePreset'
      responses:
        '200':
          description: Instance preset updated
          content:
            application/json:
              schema:
                $ref: 'crds.gen.yaml#/components/schemas/InstancePreset'
        '404':
          description: Instance preset not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - Preset
      summary: Delete instance preset
      description: |
        This API deletes the instance preset specified by the `name` in the specified `cluster`.
      operationId: deleteInstancePreset
      parameters:
        - name: cluster
          in: path
          required: true
          description: The name of the cluster
          schema:
            type: string
        - name: name
          in: path
          required: true
          description: The name of the instance preset
          schema:
            type: string
      responses:
        '204':
          description: Instance preset deleted
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  '/clusters/{cluster}/instance-presets/{name}/resolve':
    x-everest-resource-name: instance-presets
    get:
      tags:
        - Preset
      summary: Resolve instance preset with namespace defaults and default StorageClass of the cluster
      description: |
        This API resolves the instance preset specified by the `name` in the specified `cluster`.
        This endpoint is useful when creating instances from presets, as it provides a
        fully-populated preset with all namespace defaults, and default StorageClass of the cluster
        if it has empty StorageClass.
      operationId: resolveInstancePreset
      parameters:
        - name: cluster
          in: path
          required: true
          description: The name of the cluster
          schema:
            type: string
        - name: name
          in: path
          required: true
          description: The name of the instance preset
          schema:
            type: string
        - name: namespace
          in: query
          required: true
          description: The namespace for which to resolve default values
          schema:
            type: string
      responses:
        '200':
          description: Resolved instance preset with namespace defaults
          content:
            application/json:
              schema:
                $ref: 'crds.gen.yaml#/components/schemas/InstancePreset'
        '404':
          description: Instance preset not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  '/clusters/{cluster}/namespaces/{namespace}/instances/{instance}/create-preset':
    x-everest-resource-name: instance-presets
    post:
      tags:
        - Preset
      summary: Create instance preset from instance
      description: |
        This API creates a new instance preset from an existing instance.
        The preset will be populated with the instance's spec, with namespace-scoped
        fields cleared (secrets, monitoring configs, etc.).
      operationId: createInstancePresetFromInstance
      parameters:
        - name: cluster
          in: path
          required: true
          description: The name of the cluster
          schema:
            type: string
        - name: namespace
          in: path
          required: true
          description: The namespace of the instance
          schema:
            type: string
        - name: instance
          in: path
          required: true
          description: The name of the instance
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - presetName
              properties:
                presetName:
                  type: string
                  description: The name for the new preset
      responses:
        '201':
          description: Instance preset created from instance
          content:
            application/json:
              schema:
                $ref: 'crds.gen.yaml#/components/schemas/InstancePreset'
        '404':
          description: Instance not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
    Error:
      type: object
      description: Error response
      properties:
        message:
          type: string
...